Adapted the strToInt* tool functions, the fromInt*()/toInt*() member functions no tStr, and the constructors and toInt*() member functions on Var in order to gain support for int64_t in the Var.
Tested the changes by trying to compile the core and the shell (which includes
the js component). The other components have not yet been tested, so they are
probably broken after this commit.

NOTE: size_t can be assigned to Var directly. A special constructor will
automatically call through to the constructor for an int32_t (this is the
default for now. If we decided to add 64-bit support in the future, this
function should be updated to (statically) select the correct alternative)